# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Esaal medical appointment platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /doctor/profile:
    get:
      operationId: getDoctorProfile
      tags: [doctors]
      summary: Get current doctor profile
      responses:
        "200":
          description: Doctor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
    patch:
      operationId: updateDoctorProfile
      tags: [doctors]
      summary: Update current doctor profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDoctorProfileBody"
      responses:
        "200":
          description: Updated doctor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
...
  /doctors:
    get:
      operationId: getDoctors
      tags: [doctors]
      summary: Get list of doctors with optional filters
      parameters:
        - name: specialty
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [psychiatrist, psychologist]
        - name: sessionType
          in: query
          schema:
            type: string
            enum: [individual, group]
        - name: gender
          in: query
          schema:
            type: string
            enum: [male, female]
        - name: immediate
          in: query
          schema:
            type: string
        - name: freeConsultation
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"
  /doctors/{id}:
    get:
      operationId: getDoctor
      tags: [doctors]
      summary: Get a doctor's profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Doctor"
        "404":
          description: Doctor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDoctor
      tags: [doctors]
      summary: Delete a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Doctor deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /doctors/{id}/slots:
    get:
      operationId: getDoctorSlots
      tags: [slots]
      summary: Get available slots for a doctor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Slot"
...
  /doctor/dashboard:
    get:
      operationId: getDoctorDashboard
      tags: [doctors]
      summary: Doctor dashboard stats
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorDashboard"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Doctor:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        specialty:
          type: array
          items:
            type: string
        subspecialty:
          type: ["string", "null"]
        type:
          type: string
          enum: [psychiatrist, psychologist]
        sessionType:
          type: string
          enum: [individual, group, both]
        gender:
          type: string
          enum: [male, female]
        price:
          type: integer
        rating:
          type: integer
        reviewCount:
          type: integer
        bio:
          type: ["string", "null"]
        isOnline:
          type: boolean
        immediateAvailable:
          type: boolean
        freeConsultation:
          type: boolean
        yearsExperience:
          type: ["integer", "null"]
        languages:
          type: array
          items:
            type: string
        avatarUrl:
          type: ["string", "null"]
        pendingBio:
          type: ["string", "null"]
        pendingPrice:
          type: ["integer", "null"]
        pendingSpecialty:
          type: array
          items:
            type: string
        pendingLanguages:
          type: array
          items:
            type: string
        pendingGender:
          type: ["string", "null"]
        paymentInfo:
          type: ["string", "null"]
        pendingPaymentInfo:
          type: string
          nullable: true
        isRejected:
          type: boolean
        rejectionReason:
          type: string
          nullable: true
        isApproved:
          type: boolean
      required:
        - id
        - userId
        - firstName
        - lastName
        - email
        - specialty
        - type
        - sessionType
        - gender
        - price
        - rating
        - reviewCount
        - isOnline
        - immediateAvailable
        - freeConsultation
        - languages
        - isApproved
    Slot:
      type: object
      properties:
        id:
          type: integer
        doctorId:
          type: integer
        startTime:
          type: string
        endTime:
          type: string
        isBooked:
          type: boolean
      required:
        - id
        - doctorId
        - startTime
        - endTime
        - isBooked
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        doctorId:
          type: integer
        slotId:
          type: ["integer", "null"]
        isInstant:
          type: boolean
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        isPaid:
          type: boolean
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        patientFirstName:
          type: ["string", "null"]
        patientLastName:
          type: ["string", "null"]
        patientEmail:
          type: ["string", "null"]
        patientPhone:
          type: ["string", "null"]
        doctorFirstName:
          type: ["string", "null"]
        doctorLastName:
          type: ["string", "null"]
        doctorSpecialty:
          type: array
          items:
            type: string
        doctorPrice:
          type: ["integer", "null"]
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        patientRating:
          type: ["integer", "null"]
        patientReview:
          type: ["string", "null"]
        cancelledBy:
          type: ["integer", "null"]
        cancelledAt:
          type: ["string", "null"]
        doctorPaymentInfo:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - patientId
        - doctorId
        - status
        - isPaid
        - isInstant
        - createdAt
...
    DoctorDashboard:
      type: object
      properties:
        upcomingCount:
          type: integer
        completedCount:
          type: integer
        pendingPaymentCount:
          type: integer
        totalEarnings:
          type: integer
        recentAppointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
      required:
        - upcomingCount
        - completedCount
        - pendingPaymentCount
        - totalEarnings
        - recentAppointments
    UpdateDoctorProfileBody:
      type: object
      properties:
        bio:
          type: string
        isOnline:
          type: boolean
        price:
          type: integer
        specialty:
          type: array
          items:
            type: string
        paymentInfo:
          type: string
        gender:
          type: string
          enum: [male, female]
        sessionType:
          type: string
          enum: [individual, group]
        languages:
          type: array
          items:
            type: string
        yearsExperience:
          type: integer
...